fix: reserve priority keys first
Priority keys weren't actually evaluated in a priority fashion, often
leaving them unable to reserve the designated key by the time an action
was evaluated. Separate the priority/non-priority get_action_config
functions and perform a pre-pass with priority keys to ensure the
reserved keys are taken and not left up to the original actions order.